+++ iftttie/actions/telegram.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from enum import Enum
-from typing import Optional
+from typing import Any, Dict, Optional
 
 from aiohttp import ClientResponse, ClientSession
 from loguru import logger
@@ -12,6 +12,14 @@ url = 'https://api.telegram.org/bot{token}/{method}'
 class ParseMode(str, Enum):
     MARKDOWN = 'Markdown'
     HTML = 'HTML'
+
+
+async def post(session: ClientSession, token: str, method: str, data: Dict[Any, Any]) -> ClientResponse:
+    logger.info('{}', method)
+    response: ClientResponse = await session.post(url.format(token=token, method=method), data=data)
+    if response.status != 200:
+        logger.error('Telegram error: {}', await response.text())
+    return response
 
 
 async def send_message(
@@ -32,7 +40,7 @@ async def send_message(
     if parse_mode:
         data['parse_mode'] = parse_mode.value
     logger.debug('send_message(text={!r})', text)
-    return await session.post(url.format(token=token, method='sendMessage'), data=data)
+    return await post(session, token, 'sendMessage', data)
 
 
 async def send_animation(
@@ -54,4 +62,4 @@ async def send_animation(
     if parse_mode:
         data['parse_mode'] = parse_mode.value
     logger.debug('send_animation(animation={!r})', animation)
-    return await session.post(url.format(token=token, method='sendAnimation'), data=data)
+    return await post(session, token, 'sendAnimation', data)